// FrontEnd/src/interfaces/VotesModel.ts
export interface VotesModel {
  Address: string;
  Age: number;
  "Block Code": number;
  "Book No": string;
  "Boulevard|Avenue": string;
  City: string;
  Constituency: string;
  "Constituency Name": string;
  Count: number;
  Dehya: string;
  District: string;
  "Family No": number;
  "Father|Husband Name": string;
  Gender: string;
  "House No": string;
  Lane: string;
  "Marital Status": string;
  Moza: string;
  NIC: string;
  Name: string;
  "Other Area": string;
  "Patwar Halka": string;
  Phase: string;
  "S No": number;
  Sector: string;
  Street: number | string;
  Talka: string;
  Tapaydar: string;
  Tehseel: string;
  "Union Council": string;
  "Vote S No": number;
}